pythagoras/requirements.txt
Django==2.0.2
django-controlcenter==0.2.6


